+++ tests/20_get_poll_handles_test.rs
@@ -9,7 +9,8 @@ pub mod common;
 use common::{callback::MyCallbacks, param::PASSWORD, setup::TestConfiguration, utils::my_tpm_finalizer};
 use function_name::named;
 use serial_test::serial;
-use tss2_fapi_rs::{ErrorCode, FapiContext, FapiPollHandle, InternalError};
+use std::hint::black_box;
+use tss2_fapi_rs::{BaseErrorCode, ErrorCode, FapiContext, FapiPollHandle};
 
 // ==========================================================================
 // Test cases
@@ -32,24 +33,15 @@ fn test_get_poll_handles() {
         // Initialize TPM, if not already initialized
         tpm_initialize!(context, PASSWORD, MyCallbacks::new(PASSWORD, None));
 
-        // Try to acquire the poll handles (this is expected to fail!)
+        // Try to acquire the poll handles
         let result: Result<Vec<FapiPollHandle>, _> = context.get_poll_handles();
-        assert!(matches!(result, Err(ErrorCode::InternalError(InternalError::NotImplemented))));
 
-        // Check the handles
-        if let Ok(poll_handles) = result {
-            for handle in poll_handles {
-                check_poll_handle(&handle);
-            }
+        // Expect `BadSequence` error here, because poll handles are only available while in the middle of an asynchronious I/O operation!
+        assert!(matches!(result, Err(ErrorCode::FapiError(BaseErrorCode::BadSequence))));
+
+        // Verify handles
+        if let Ok(handles) = result {
+            handles.into_iter().for_each(|handle| _ = black_box(handle.0));
         }
     });
-}
-
-// ==========================================================================
-// Helper functions
-// ==========================================================================
-
-#[inline(always)]
-fn check_poll_handle(handle: &FapiPollHandle) {
-    _ = std::hint::black_box(handle);
 }